+++ crates/biome_js_analyze/tests/specs/nursery/noRestrictedImports/valid.js
@@ -1,1 +1,2 @@
 const path = require('lodash');
+const path = require('lodash/get');