+++ crates/biome_js_analyze/tests/specs/nursery/noRestrictedImports/valid.ts
@@ -1,1 +1,3 @@
 declare module "node:fs" {}
+declare module "lodash/get" {}
+declare module "lodash/sortBy" {}