+++ CHANGELOG.md
@@ -52,6 +52,7 @@ our [guidelines for writing a good changelog entry](https://github.com/biomejs/b
 - Add [noImgElement](https://biomejs.dev/linter/rules/no-img-element/). Contributed by @kaioduarte
 - Add [guardForIn](https://biomejs.dev/linter/rules/guard-for-in/). Contributed by @fireairforce
 - Add [noUselessStringRaw](https://github.com/biomejs/biome/pull/4263). Contributed by @fireairforce
+- Add an option `patterns` to [noRestrictedImports](https://biomejs.dev/linter/rules/no-restricted-imports/). Contributed by @nekocode
 
 #### Bug Fixes